backend/requirements.txt
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
# mysqlclient==2.2.0
drf-yasg==1.21.7
